routes/controllers

+++ database/migrations/2023_04_19_235217_create_products_table.php
@@ -16,7 +16,7 @@ class CreateProductsTable extends Migration
         Schema::create('products', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->integer('category_id');
+            $table->unsignedBigInteger('category_id');
             $table->text('detail');
             $table->timestamps();
